requirements.txt
asgiref==3.4.1
cffi==1.15.1
cryptography==37.0.4
Django==3.2.14
django-autoslug==1.9.8
django-cors-headers==3.10.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
Pillow==8.4.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==1.7.1
python-http-client==3.3.7
pytz==2022.1
sendgrid==6.9.7
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
typing_extensions==4.1.1
